\name{slacR-package}
\alias{slacR-package}
\alias{slacR}
\docType{package}
\title{
  Kinematic analysis of galaxy spectra
}
\description{
  This package analyses galaxy spectra to yield measurements of
  systematic velocity and velocity dispersion. The line-of-sight
  velocity distribution is assumed to be Gaussian, higher moments are
  not measured. 
}
\details{
  \tabular{ll}{
    Package: \tab slacR\cr
    Type: \tab Package\cr
    Version: \tab 3.1.5\cr
    Date: \tab 2014-04-30\cr
    License: \tab GPL\cr
  }
}
\author{
  Oliver Czoske
  
  Maintainer: Oliver Czoske <oliver.czoske@univie.ac.at>
}
\references{
}
\keyword{package}
